Changes in H2PInteractions. Serious changes.
- new way Antigen works needs a new way all that stuff works.

+++ Gene.cpp
@@ -237,7 +237,7 @@ genestring Gene::getBitGene(){
  *
  * @return integer representation of a gene.
  */
-int Gene::getTheRealGene(){
+unsigned long int Gene::getTheRealGene(){
     return TheGene;
 }
